x86_64 patch from Bernd
from: eda03ea1f7554d8bc24c6443b7a405d0ed84e9a0

+++ lib/tas/src/lua/llimits.h
@@ -41,7 +41,7 @@
 */
 
 /* an unsigned integer to hold hash values */
-typedef unsigned int lu_hash;
+typedef unsigned long int lu_hash;
 /* its signed equivalent */
 typedef int ls_hash;
